修复告警

+++ frameworks/core/components_ng/pattern/text_field/text_field_pattern.cpp
@@ -3693,7 +3693,7 @@ bool TextFieldPattern::FilterWithAscii(const std::string& valueToUpdate, std::st
     }
     bool textChange = true;
     std::string errorText = "";
-    for (uint32_t valuePtr = 0; valuePtr < valueToUpdate.size(); valuePtr++) {
+    for (size_t valuePtr = 0; valuePtr < valueToUpdate.size(); valuePtr++) {
         if (isascii(valueToUpdate[valuePtr])) {
             result += valueToUpdate[valuePtr];
         } else {
